% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_queries.R
\name{absolute_form_query}
\alias{absolute_form_query}
\title{Create form query of type 'ABSOLUTE'}
\usage{
absolute_form_query(
  query_id,
  start_date,
  end_date,
  queries = NULL,
  concept_ids = NULL,
  connection = NULL,
  resolution = "COMPLETE"
)
}
\arguments{
\item{query_id}{query_id for query that defines population}

\item{start_date}{start date of date restriction for features}

\item{end_date}{end date of date restriction for features}

\item{queries}{list of queries
(type 'CONCEPT_QUERY' - from create_query function)}

\item{concept_ids}{list of concept ids to use instead of queries}

\item{connection}{connection object is needed when using concept_ids}

\item{resolution}{time resolution of output - 'COMPLETE', 'YEARS', 'QUARTERS'}
}
\value{
absolute form query
}
\description{
Creates query of type 'ABSOLUTE'
}
\examples{

# Step 1. Get query id that represents group of people
# a) Take query id from return value of execute_query function (Example)
concept_id = "fdb_destatis.alter"
query = concept_to_query(concept_id, connection = connection,
                         start_date="2017-01-01", end_date="2017-01-31")
query_id = execute_query(connection, query)

# b) Get query id from from query label
query_id = query_label_to_id(connection, "query-label")

# Step 2. Create Form Query

## Absolute Form
form_query = absolute_form_query(query_id,
                                 concept_ids = list("fdb_destatis.alter",
                                                    "fdb_destatis.geschlecht"),
                                 connection=connection,
                                 start_date="2017-01-01", end_date="2017-01-31")

## Relative Form
form_query = relative_form_query(query_id,
                                 before_concept_ids = list("fdb_destatis.alter",
                                                    "fdb_destatis.geschlecht"),
                                 after_concept_ids = list("fdb_destatis.alter",
                                                           "fdb_destatis.geschlecht"),
                                 connection=connection,
                                 start_date="2017-01-01", end_date="2017-01-31")

# Step 3. Execute form_query
form_query_id = execute_query(connection, form_query)

# Step 4. Get query result
data = get_query_result(connection, form_query_id)
}
