% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{change_dataset}
\alias{change_dataset}
\title{Change dataset}
\usage{
change_dataset(connection, dataset)
}
\arguments{
\item{connection}{connection obj}

\item{dataset}{dataset to access}
}
\value{
connection object
}
\description{
Changes dataset in connection to access different dataset
}
