% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{concept_id_to_query}
\alias{concept_id_to_query}
\title{Creates query from concept id}
\usage{
concept_id_to_query(connection, concept_id, start_date = NULL, end_date = NULL)
}
\arguments{
\item{connection}{connection object}

\item{concept_id}{concept id of concept to turn into query}

\item{start_date}{start date, if date restriction is wanted}

\item{end_date}{end date, if date restriction is wanted}
}
\value{
query
}
\description{
Takes concept id and connection to create query from concept
A date range can be added.
}
\examples{

# Create and execute simple query

# Step 1. Define concept to use
#         (you can look up all possible ids using get_concepts())
concept_id = "adb_energie.alter"

# Step 2. Create query from concept id and store query in variable
#         Similar to dragging concept in editor field in frontend
query = concept_to_query(concept_id, connection = connection, start_date="2017-01-01", end_date="2017-01-31")

# Step 3. Execute query you created and store query id in variable
query_id = execute_query(connection, query)

# Step 4. Get query result (by default a data.table format is returned)
data = get_query_result(connection, query_id)
}
