% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_queries.R
\name{concept_ids_to_concept_objs}
\alias{concept_ids_to_concept_objs}
\title{Creates concept objects from concept ids to use in export form}
\usage{
concept_ids_to_concept_objs(connection, concept_ids)
}
\description{
Creates concept objects from concept ids to use in export form
}
