% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{execute_query}
\alias{execute_query}
\title{Execute query}
\usage{
execute_query(connection, query)
}
\arguments{
\item{connection}{connection object}

\item{query}{query that will be executed}
}
\value{
query
}
\description{
Executes query and returns query_id
}
