% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_api_functions.R
\name{get}
\alias{get}
\title{get request for conquery}
\usage{
get(connection)
}
\arguments{
\item{connection}{connection object}
}
\value{
response of request
}
\description{
makes get request with credentials and url from connection
}
