% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_concept}
\alias{get_concept}
\title{Get concept from conquery}
\usage{
get_concept(connection, concept_id)
}
\arguments{
\item{connection}{connection object}

\item{concept_id}{concept id (always includes -dataset- in front,
which has to match dataset in connection)}
}
\value{
concept object
}
\description{
Returns concept with given concept_id
}
