% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_concepts}
\alias{get_concepts}
\title{Get concepts from conquery}
\usage{
get_concepts(connection, include_struc_elements = FALSE)
}
\arguments{
\item{connection}{connection object}

\item{include_struc_elements}{when TRUE, structure elements of concepts
in frontend are included in concept list}
}
\value{
concepts of specific dataset in conquery
}
\description{
Returns list with concepts of dataset specified in connection
}
\examples{

# load concepts of dataset
concepts <- get_concepts(connection)
}
