% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_connection}
\alias{get_connection}
\title{Creates connection object}
\usage{
get_connection(dataset, url = NULL, token = NULL)
}
\arguments{
\item{dataset}{dataset to access}

\item{url}{url to conquery}

\item{token}{token to access dataset (use login function to add credentials)}
}
\value{
connection object
}
\description{
The connection object stores url to conquery, credentials and dataset
}
\examples{

# Step 1. Create a connection
connection <- get_connection("fdb_destatis")

# Step 2. Login with your username and password
## Normal login
connection <- login(connection)

## Login with predefined username
connection <- login(connection, user = "max.knobloch@ingef.de")

# Step 3. Validate login and permission on dataset
validate_connection(connection)

# Change dataset of connection
connection <- change_dataset(connection, "fdb_demo")

}
