% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_query}
\alias{get_query}
\title{Get query from query_id}
\usage{
get_query(connection, query_id)
}
\arguments{
\item{connection}{connection object}

\item{query_id}{query id (always includes -dataset- in front,
which has to match dataset in connection)}
}
\value{
query
}
\description{
Returns query of query with given query_id
}
