% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_query_info}
\alias{get_query_info}
\title{Get query information from query_id}
\usage{
get_query_info(connection, query_id)
}
\arguments{
\item{connection}{connection object}

\item{query_id}{query id (always includes -dataset- in front,
which has to match dataset in connection)}
}
\value{
query information
}
\description{
Returns query information of query with given query_id
}
