% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_query_result}
\alias{get_query_result}
\title{Get result of executed query}
\usage{
get_query_result(connection, query_id, data_format = "data.table")
}
\arguments{
\item{connection}{connection object}

\item{query_id}{query_id of the executed query}

\item{data_format}{when 'data.table' it returns data in a data.table.
If 'raw', data will be returns as csv text}
}
\value{
query result
}
\description{
Returns data of executed query in data.table or raw text format
}
\examples{

# Create and execute simple query

# Step 1. Define concept to use
#         (you can look up all possible ids using get_concepts())
concept_id = "adb_energie.alter"

# Step 2. Create query from concept id and store query in variable
#         Similar to dragging concept in editor field in frontend
query = concept_to_query(concept_id, connection = connection, start_date="2017-01-01", end_date="2017-01-31")

# Step 3. Execute query you created and store query id in variable
query_id = execute_query(connection, query)

# Step 4. Get query result (by default a data.table format is returned)
data = get_query_result(connection, query_id)
}
