% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_stored_queries}
\alias{get_stored_queries}
\title{Get all stored queries}
\usage{
get_stored_queries(connection)
}
\arguments{
\item{connection}{connection object}
}
\value{
list of stored queries
}
\description{
Returns all stored queries that are save on the dataset in conncetion
}
