% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{login}
\alias{login}
\title{Log in with log in data}
\usage{
login(connection, user = NULL)
}
\arguments{
\item{connection}{connection object}
}
\value{
connection object with credential
}
\description{
log in with user name and password to update connection
}
