% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_api_functions.R
\name{post}
\alias{post}
\title{post request for conquery}
\usage{
post(connection, json_data)
}
\arguments{
\item{connection}{connection object}

\item{json_data}{data to send}
}
\value{
response of request
}
\description{
makes post request with credentials and url from connection
}
