% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{query_label_to_id}
\alias{query_label_to_id}
\title{Matches query label to query id}
\usage{
query_label_to_id(connection, query_label)
}
\arguments{
\item{connection}{connection object}

\item{query_label}{label to query}
}
\value{
query id that belongs to label
}
\description{
Looks for label in all stored queries and returns query id if label is found
}
