% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{table_export_query}
\alias{table_export_query}
\title{Create table export query}
\usage{
table_export_query(query, start_date, end_date, connector_id, date_column)
}
\arguments{
\item{query}{query that defines population}

\item{start_date}{start date for date restriction of table data}

\item{end_date}{end date for date restriction of table data}

\item{connector_id}{id of connector that contains raw data}

\item{date_column}{column name of column that is checked for date restriction}
}
\value{
table export query
}
\description{
Creates table export query for getting raw data in conquery
}
\examples{
table_query = table_export_query(query=get_query(connection, query_id),
                                 start_date="2017-01-01",
                                 end_date="2017-01-31",
                                 connector_id = "adb_energie.alter.alter",
                                 date_column = "adb_energie.alter.alter.versichertenzeit")
table_query_id = execute_query(connection, table_query)
}
