% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{validate_connection}
\alias{validate_connection}
\title{Validates connection}
\usage{
validate_connection(connection)
}
\arguments{
\item{connection}{connection obj}

\item{dataset}{dataset to access}
}
\description{
Checks if there is permission to access any dataset
}
