% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{concept_to_query}
\alias{concept_to_query}
\title{Creates query from concept id}
\usage{
concept_to_query(
  concept_id,
  connection = NULL,
  concept = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{concept_id}{concept id of concept to turn into query}

\item{connection}{connection object (only necessary when concept is NULL)}

\item{concept}{concept that is turned into query}

\item{start_date}{start date, if date restriction is wanted}

\item{end_date}{end date, if date restriction is wanted}
}
\value{
query
}
\description{
Takes either concept (from concepts list - get_concepts function) or takes
connection to get the concept by itself and returns a query.
A date range can be added.
}
