% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_query_result}
\alias{get_query_result}
\title{Get result of executed query}
\usage{
get_query_result(connection, query_id, data_format = "data.table")
}
\arguments{
\item{connection}{connection object}

\item{query_id}{query_id of the executed query}

\item{data_format}{when 'data.table' it returns data in a data.table.
If 'raw', data will be returns as csv text}
}
\value{
query result
}
\description{
Returns data of executed query in data.table or raw text format
}
